Style file added for cleaner code and reutilizable components
